requirements.txt
Flask==0.10.1
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
nested-dict==1.61
numpy==1.11.0
pandas==0.18.0
python-dateutil==2.5.2
python-Levenshtein==0.12.0
pytz==2016.3
pyxDamerauLevenshtein==1.3.2
six==1.10.0
Werkzeug==0.11.5
wget==3.2
